pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nde-analysis"
version = "0.1.0"
description = "CLI analysis pipeline for NDE publication workflows"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "scipy>=1.12",
  "statsmodels>=0.14",
  "scikit-learn>=1.4",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
]

[project.scripts]
nde-analysis = "nde_analysis.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
